requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.7.0
tqdm>=4.66.0
numpy>=1.26.0
sentence-transformers>=2.6.1
sqlite-vec>=0.1.6
json5>=0.9.25


